try3
with new customer FXML and Controller
