+++ src/main/java/com/example/fdpuser3/OnlineCustomerController.java
@@ -1,0 +1,220 @@
+package com.example.fdpuser3;
+
+import javafx.event.ActionEvent;
+import javafx.scene.chart.BarChart;
+import javafx.scene.chart.LineChart;
+import javafx.scene.control.*;
+import javafx.scene.layout.AnchorPane;
+
+public class OnlineCustomerController {
+    @javafx.fxml.FXML
+    private TableView tableviewCancelOrderPane;
+    @javafx.fxml.FXML
+    private AnchorPane cancelOrdersPane;
+    @javafx.fxml.FXML
+    private TableView viewNoticeTable;
+    @javafx.fxml.FXML
+    private Label totalPriceLabelId1;
+    @javafx.fxml.FXML
+    private CheckBox bagReqForEquipPaneCheckBox;
+    @javafx.fxml.FXML
+    private AnchorPane requestForEqupPane;
+    @javafx.fxml.FXML
+    private TableColumn deliveryColHrCancelOrderPane;
+    @javafx.fxml.FXML
+    private Label requestForEquipLabel;
+    @javafx.fxml.FXML
+    private TableColumn foodNameColId1;
+    @javafx.fxml.FXML
+    private DatePicker startDateIncomeReport;
+    @javafx.fxml.FXML
+    private Label viewNoticeLabel;
+    @javafx.fxml.FXML
+    private AnchorPane incomePane;
+    @javafx.fxml.FXML
+    private ComboBox selectPickUpHourComboBox;
+    @javafx.fxml.FXML
+    private AnchorPane submitApplicationPane;
+    @javafx.fxml.FXML
+    private AnchorPane helpPane;
+    @javafx.fxml.FXML
+    private TableColumn subTotalColId1;
+    @javafx.fxml.FXML
+    private TableView tableViewId1;
+    @javafx.fxml.FXML
+    private LineChart lineChartIncomeReportId;
+    @javafx.fxml.FXML
+    private TableColumn pickUpHrColId;
+    @javafx.fxml.FXML
+    private AnchorPane OrderPane;
+    @javafx.fxml.FXML
+    private TableColumn foodIdCol;
+    @javafx.fxml.FXML
+    private TableColumn foodNameColId;
+    @javafx.fxml.FXML
+    private Label totalPriceLabelId;
+    @javafx.fxml.FXML
+    private DatePicker endDateIncomeReport;
+    @javafx.fxml.FXML
+    private TableColumn dateColViewNoticeTable;
+    @javafx.fxml.FXML
+    private TableColumn quantityColId;
+    @javafx.fxml.FXML
+    private TableView tableViewId;
+    @javafx.fxml.FXML
+    private BarChart barChartIncomeReportId;
+    @javafx.fxml.FXML
+    private ComboBox pairOfGlovesReqForEquipPaneComboBox;
+    @javafx.fxml.FXML
+    private TableColumn subjectColViewNoticeTable;
+    @javafx.fxml.FXML
+    private TableColumn deliveryTypeColId1;
+    @javafx.fxml.FXML
+    private TableColumn deliveryMinColCancelOrderPane;
+    @javafx.fxml.FXML
+    private TableColumn foodIdCol1;
+    @javafx.fxml.FXML
+    private ComboBox selectFoodTypeComboBox;
+    @javafx.fxml.FXML
+    private CheckBox helmetReqForEquipPaneCheckBox;
+    @javafx.fxml.FXML
+    private TextField textFieldSubmitButton;
+    @javafx.fxml.FXML
+    private TableColumn pickUpHrColId1;
+    @javafx.fxml.FXML
+    private TableColumn deliveryTypeColId;
+    @javafx.fxml.FXML
+    private TableColumn addressColCancelOrderPane;
+    @javafx.fxml.FXML
+    private Label descriptionOfGoalsLabelHelpPane;
+    @javafx.fxml.FXML
+    private TableColumn DeliveryChargeColCancelOrderPane;
+    @javafx.fxml.FXML
+    private AnchorPane viewNoticePane;
+    @javafx.fxml.FXML
+    private ComboBox selectPickUpMinComboBox1;
+    @javafx.fxml.FXML
+    private TableColumn subTotalColId;
+    @javafx.fxml.FXML
+    private ComboBox selectFoodTypeComboBox1;
+    @javafx.fxml.FXML
+    private CheckBox phoneCarrierReqForEquipPaneCheckBox;
+    @javafx.fxml.FXML
+    private ComboBox selectQuantityComboBox1;
+    @javafx.fxml.FXML
+    private TableColumn quantityColId1;
+    @javafx.fxml.FXML
+    private TableColumn packageColCancelOrderPane;
+    @javafx.fxml.FXML
+    private AnchorPane pickUpPane;
+    @javafx.fxml.FXML
+    private TableColumn typeColViewNoticeTable;
+    @javafx.fxml.FXML
+    private ComboBox selectQuantityComboBox;
+    @javafx.fxml.FXML
+    private CheckBox TShirtReqForEquipPaneCheckBox;
+
+    @javafx.fxml.FXML
+    public void createPDF_IncomeReportButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void orderFoodButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void backButon(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void reportButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void tipGivingButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void helpButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howToRenewContract_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void cancelOrderButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void rateFoodButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void placeOrderButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void sendButtonSubmitAApplication(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howToReqForEquip_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void viewIntoTableButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howToViewNotice_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void pickUpButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howToSubmitApplication_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void orderHistory(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howToCancelOrders_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void loadButtonReqForEquipPaneCheckBox(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void CancelButtonCheckorderPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void barChartIncomeReportButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void LineChartIncomeReportButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void pickUpOrderButton(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void howTo_CheckOrders_ButtonHelpPane(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void submitButtonReqForEquipPaneCheckBox(ActionEvent actionEvent) {
+    }
+
+    @javafx.fxml.FXML
+    public void delayButtonCheckorderPane(ActionEvent actionEvent) {
+    }
+}
