--- src/main/java/com/example/fdpuser3/fdpuser3/AppendableObjectOutputStream.java
@@ -1,21 +1,0 @@
-package com.example.fdpuser3.fdpuser3;
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
-
-
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.io.OutputStream;
-
-public class AppendableObjectOutputStream extends ObjectOutputStream {
-    public AppendableObjectOutputStream(OutputStream out) throws IOException {
-        super(out);
-    }
-
-    protected void writeStreamHeader() throws IOException {
-    }
-}
-
